+++ models/model_neuron_py/model_neuron_py.h
@@ -30,16 +30,16 @@ extern const std::array<std::string_view const, 0> expression_ids;
 extern std::array<int, 2> state_idxs_solver;
 
 extern void Jy_model_neuron_py(realtype *Jy, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my);
-extern void dJydsigma_model_neuron_py(realtype *dJydsigma, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my);
+extern void dJydsigmay_model_neuron_py(realtype *dJydsigmay, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my);
 extern void dJydy_model_neuron_py(realtype *dJydy, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my);
 extern void dJydy_colptrs_model_neuron_py(SUNMatrixWrapper &colptrs, int index);
 extern void dJydy_rowvals_model_neuron_py(SUNMatrixWrapper &rowvals, int index);
 extern void Jz_model_neuron_py(realtype *Jz, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const realtype *mz);
-extern void dJzdsigma_model_neuron_py(realtype *dJzdsigma, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const realtype *mz);
+extern void dJzdsigmaz_model_neuron_py(realtype *dJzdsigmaz, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const realtype *mz);
 extern void dJzdz_model_neuron_py(realtype *dJzdz, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const double *mz);
 extern void Jrz_model_neuron_py(realtype *Jrz, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz);
-extern void dJrzdsigma_model_neuron_py(realtype *dJrzdsigma, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz);
-extern void dJrzdz_model_neuron_py(realtype *dJrzdz, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz);
+extern void dJrzdsigmaz_model_neuron_py(realtype *dJrzdsigmaz, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz);
+extern void dJrzdrz_model_neuron_py(realtype *dJrzdrz, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz);
 extern void root_model_neuron_py(realtype *root, const realtype t, const realtype *x, const realtype *p, const realtype *k, const realtype *h, const realtype *w, const realtype *tcl);
 
 
@@ -180,23 +180,23 @@ class Model_model_neuron_py : public amici::Model_ODE {
     }
 
 
-    void fdJrzdsigma(realtype *dJrzdsigma, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz) override {
-        dJrzdsigma_model_neuron_py(dJrzdsigma, iz, p, k, rz, sigmaz);
+    void fdJrzdsigmaz(realtype *dJrzdsigmaz, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz) override {
+        dJrzdsigmaz_model_neuron_py(dJrzdsigmaz, iz, p, k, rz, sigmaz);
     }
 
 
-    void fdJrzdz(realtype *dJrzdz, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz) override {
-        dJrzdz_model_neuron_py(dJrzdz, iz, p, k, rz, sigmaz);
+    void fdJrzdrz(realtype *dJrzdrz, const int iz, const realtype *p, const realtype *k, const realtype *rz, const realtype *sigmaz) override {
+        dJrzdrz_model_neuron_py(dJrzdrz, iz, p, k, rz, sigmaz);
     }
 
 
-    void fdJydsigma(realtype *dJydsigma, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my) override {
-        dJydsigma_model_neuron_py(dJydsigma, iy, p, k, y, sigmay, my);
+    void fdJydsigmay(realtype *dJydsigmay, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my) override {
+        dJydsigmay_model_neuron_py(dJydsigmay, iy, p, k, y, sigmay, my);
     }
 
 
-    void fdJzdsigma(realtype *dJzdsigma, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const realtype *mz) override {
-        dJzdsigma_model_neuron_py(dJzdsigma, iz, p, k, z, sigmaz, mz);
+    void fdJzdsigmaz(realtype *dJzdsigmaz, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const realtype *mz) override {
+        dJzdsigmaz_model_neuron_py(dJzdsigmaz, iz, p, k, z, sigmaz, mz);
     }
 
 
@@ -558,7 +558,7 @@ class Model_model_neuron_py : public amici::Model_ODE {
      * @return AMICI git commit hash
      */
     std::string get_amici_commit() const override {
-        return "f04344fe181ed0b63917c93672370d76e427fb22";
+        return "088c23bd7aa5411508ed82dc4f1ed8b565a69713";
     }
 
     bool has_quadratic_llh() const override {